+++ backend/src/main/java/com/techeerlog/global/mapper/ProjectMapper.java
@@ -3,8 +3,19 @@ package com.techeerlog.global.mapper;
 import com.techeerlog.framework.dto.FrameworkResponse;
 import com.techeerlog.member.dto.MemberResponse;
 import com.techeerlog.project.domain.Project;
-import com.techeerlog.project.dto.*;
-import org.mapstruct.*;
+
+import com.techeerlog.project.dto.response.NonRegisterProjectMemberResponse;
+import com.techeerlog.project.dto.response.ProjectItemResponse;
+import com.techeerlog.project.dto.response.ProjectMemberResponse;
+import com.techeerlog.project.dto.request.ProjectRequest;
+import com.techeerlog.project.dto.response.ProjectResponse;
+import org.mapstruct.BeanMapping;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.MappingTarget;
+import org.mapstruct.NullValuePropertyMappingStrategy;
+import org.mapstruct.ReportingPolicy;
+
 
 import java.util.List;
