+++ backend/src/main/java/com/techeerlog/project/dto/request/ProjectRequest.java
@@ -1,7 +1,9 @@
-package com.techeerlog.project.dto;
+package com.techeerlog.project.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.techeerlog.framework.dto.FrameworkRequest;
+import com.techeerlog.project.dto.request.NonRegisterProjectMemberRequest;
+import com.techeerlog.project.dto.request.ProjectMemberRequest;
 import com.techeerlog.project.enums.PlatformEnum;
 import com.techeerlog.project.enums.ProjectStatusEnum;
 import com.techeerlog.project.enums.ProjectTypeEnum;
