+++ backend/src/main/java/com/techeerlog/project/dto/response/NonRegisterProjectMemberResponse.java
@@ -1,16 +1,17 @@
-package com.techeerlog.project.dto;
+package com.techeerlog.project.dto.response;
 
 import com.techeerlog.project.enums.ProjectMemberTypeEnum;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Getter
-@Setter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class NonRegisterProjectMemberRequest {
+public class NonRegisterProjectMemberResponse {
     private String name;
     private ProjectMemberTypeEnum projectMemberTypeEnum;
 }
