+++ backend/src/main/java/com/techeerlog/project/dto/response/ProjectItemListResponse.java
@@ -1,4 +1,4 @@
-package com.techeerlog.project.dto;
+package com.techeerlog.project.dto.response;
 
 import lombok.*;
