+++ backend/src/main/java/com/techeerlog/project/dto/response/ProjectItemResponse.java
@@ -1,4 +1,4 @@
-package com.techeerlog.project.dto;
+package com.techeerlog.project.dto.response;
 
 import com.techeerlog.framework.dto.FrameworkResponse;
 import com.techeerlog.member.dto.MemberResponse;
@@ -13,11 +13,10 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Getter
-@Setter // deserializer 시 필요함
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ProjectResponse {
+public class ProjectItemResponse {
     private Long id;
     private String mainImageUrl;
     private String title;
@@ -30,6 +29,7 @@ public class ProjectResponse {
     private ProjectTeamNameEnum projectTeamNameEnum;
     private int year;
     private SemesterEnum semesterEnum;
+    private RankEnum rankEnum;
     private ProjectStatusEnum projectStatusEnum;
     private String githubLink;
     private String blogLink;
@@ -38,7 +38,5 @@ public class ProjectResponse {
     private boolean isLoved;
     private boolean isScraped;
     private MemberResponse writer;
-    private List<ProjectMemberResponse> projectMemberResponseList;
-    private List<NonRegisterProjectMemberResponse> nonRegisterProjectMemberResponseList;
     private List<FrameworkResponse> frameworkResponseList;
 }
