+++ backend/src/main/java/com/techeerlog/project/service/ProjectService.java
@@ -20,16 +20,16 @@ import com.techeerlog.project.domain.NonRegisterProjectMember;
 import com.techeerlog.project.domain.Project;
 import com.techeerlog.project.domain.ProjectFramework;
 import com.techeerlog.project.domain.ProjectMember;
-import com.techeerlog.project.dto.NonRegisterProjectMemberRequest;
-import com.techeerlog.project.dto.NonRegisterProjectMemberResponse;
-import com.techeerlog.project.dto.PrizeProjectListRequest;
-import com.techeerlog.project.dto.ProjectItemListResponse;
-import com.techeerlog.project.dto.ProjectItemResponse;
-import com.techeerlog.project.dto.ProjectListRequest;
-import com.techeerlog.project.dto.ProjectMemberRequest;
-import com.techeerlog.project.dto.ProjectMemberResponse;
-import com.techeerlog.project.dto.ProjectRequest;
-import com.techeerlog.project.dto.ProjectResponse;
+import com.techeerlog.project.dto.request.NonRegisterProjectMemberRequest;
+import com.techeerlog.project.dto.response.NonRegisterProjectMemberResponse;
+import com.techeerlog.project.dto.request.PrizeProjectListRequest;
+import com.techeerlog.project.dto.response.ProjectItemListResponse;
+import com.techeerlog.project.dto.response.ProjectItemResponse;
+import com.techeerlog.project.dto.request.ProjectListRequest;
+import com.techeerlog.project.dto.request.ProjectMemberRequest;
+import com.techeerlog.project.dto.response.ProjectMemberResponse;
+import com.techeerlog.project.dto.request.ProjectRequest;
+import com.techeerlog.project.dto.response.ProjectResponse;
 import com.techeerlog.project.enums.RankEnum;
 import com.techeerlog.project.enums.SearchFieldEnum;
 import com.techeerlog.project.enums.SemesterEnum;
